refactor: Rename TextureReferences => TextureDatabase and extract single file classes to their own file.
