+++ Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Data/src/TextureId.cs
@@ -1,0 +1,2 @@
+namespace CaptainCoder.Dungeoneering.Unity.Data;
+public record struct TextureId(int Id);
