+++ Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Data/src/TextureId.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 60f0f10d35032d84d8097e031a5b9939
