+++ Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Data/src/TextureReference.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d595f2e9f735a5e458a522a83fec9288
