+++ Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Data/src/TileReference.cs
@@ -1,0 +1,3 @@
+using CaptainCoder.Dungeoneering.DungeonMap;
+namespace CaptainCoder.Dungeoneering.Unity.Data;
+public record struct TileReference(Dungeon Dungeon, Position Position);
