+++ Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Data/src/WallReference.cs
@@ -1,0 +1,3 @@
+using CaptainCoder.Dungeoneering.DungeonMap;
+namespace CaptainCoder.Dungeoneering.Unity.Data;
+public record struct WallReference(Dungeon Dungeon, Position Position, Facing Facing);
